refactor: split format.rs into annotate, spacing, and format modules
- Extract TokenAnnotation enum, keyword lists, and annotate_tokens into annotate.rs
- Extract PrevTokenKind, ColonContext, SpacingState, and maybe_space into spacing.rs
- Move annotation unit tests to separate annotate_tests.rs file
- Trim format.rs to only format-string generation logic
