+++ macros/src/parse/format.rs
@@ -1,669 +1,11 @@
 use proc_macro2::{Delimiter, Spacing, TokenStream, TokenTree};
 
+use super::annotate::{
+    CONTROL_FLOW_KEYWORDS, DECLARATION_KEYWORDS, TokenAnnotation, annotate_tokens,
+};
+use super::spacing::{ColonContext, PrevTokenKind, SpacingState, maybe_space};
 use super::types::{CompileError, InterpolationKind, MacroLang, TypedArg};
 use super::util::is_ident;
-
-/// Annotations computed by pre-scanning the token stream.
-/// Indexed by token position — each token gets one annotation.
-#[derive(Clone, Copy, Debug, PartialEq, Default)]
-pub(super) enum TokenAnnotation {
-    #[default]
-    Normal,
-    /// Second `:` of `::` — suppress space after it.
-    PathSepComplete,
-    /// `<` used as generic opener (matched via stack).
-    GenericOpen,
-    /// `>` used as generic closer (matched via stack).
-    GenericClose,
-    /// `&` or `*` used as prefix operator (not binary).
-    PrefixOp,
-    /// `!` used as macro-call bang (after ident).
-    MacroBang,
-    /// `?` in `?.` safe-call — suppress space before it.
-    SafeCallQ,
-    /// First `+` of `++` or first `-` of `--` used as postfix — suppress space before.
-    PostfixIncDec,
-    /// `*` used as postfix pointer marker (e.g. `Config*`) — suppress space before.
-    PostfixStar,
-    /// `?` used as postfix type marker (e.g. `Int?`, `String?`) — suppress space before.
-    PostfixQuestion,
-    /// `-` starting `->` when adjacent to preceding token (member access, not type arrow).
-    ArrowOp,
-    /// First `:` of `::` used as operator (not path separator) — space before it.
-    DoubleColonOp,
-    /// Group open (paren/bracket) that is span-adjacent to preceding token —
-    /// suppress space (function call, array index).
-    CallOpen,
-    /// `=` span-adjacent to preceding token (shell-style `NAME=val`) — suppress space.
-    AssignAdjacent,
-    /// `:` span-adjacent to both neighbors (Lua method call `obj:method()`).
-    MethodCallColon,
-    /// `-` that acts as flag prefix (like `-q`, `-f`). Does NOT suppress space
-    /// before (so `declare -q` keeps the space). Only suppresses space after via
-    /// `PrevTokenKind::PrefixOp`.
-    DashFlag,
-    /// `-` span-adjacent to both neighbors where prev is ident/literal
-    /// (hyphenated word like `from-oci-layout`).
-    DashSep,
-    /// `/` span-adjacent to both neighbors (path separator like `linux/amd64`).
-    SlashSep,
-    /// `.` used as standalone argument in shell (e.g. `find .`), not member access.
-    DotArg,
-}
-
-/// What kind of token was just emitted (for spacing decisions).
-#[derive(Clone, Copy, PartialEq)]
-pub(super) enum PrevTokenKind {
-    None,
-    Ident,
-    TypeIdent,
-    Keyword,
-    Punct(char, Spacing),
-    PrefixOp(char),
-    PathSep,
-    GenericOpen,
-    Literal,
-    GroupOpen,
-    Specifier,
-    /// `%W` soft-break — already provides a space, so suppress `maybe_space`.
-    SoftBreak,
-    /// `$$` literal dollar — suppress space after it so `$$1` renders as `$1`.
-    DollarLiteral,
-}
-
-/// Context for how `:` should be spaced.
-#[derive(Clone, Copy, PartialEq)]
-pub(super) enum ColonContext {
-    /// `name: Type`, `param: Type` — no space before `:`.
-    TypeAnnotation,
-    /// `key: value` in map/object literals — no space before `:`.
-    MapEntry,
-    /// `cond ? a : b` — space before `:`.
-    Ternary,
-    /// `std::mem` — no space before `:`.
-    PathSeparator,
-    /// `x := 42` — space before `:`.
-    WalrusAssign,
-    /// `for (item : collection)` — space before `:`.
-    ForRange,
-}
-
-/// Accumulated state threaded through the format-string builder.
-pub(super) struct SpacingState {
-    pub prev: PrevTokenKind,
-    pub colon_ctx: ColonContext,
-    /// End position (line, column) of the last specifier's closing group,
-    /// used to detect adjacent specifiers like `$L("a")$L("b")`.
-    pub prev_specifier_end: Option<(usize, usize)>,
-}
-
-impl SpacingState {
-    pub fn new() -> Self {
-        Self {
-            prev: PrevTokenKind::None,
-            colon_ctx: ColonContext::TypeAnnotation,
-            prev_specifier_end: None,
-        }
-    }
-}
-
-#[rustfmt::skip]
-pub(super) const CONTROL_FLOW_KEYWORDS: &[&str] = &[
-    "if", "else", "for", "while", "do", "switch", "catch",
-    "synchronized", "when", "guard", "unless", "until",
-    "elif", "elsif", "match", "case", "try", "with",
-    "return", "throw", "yield", "await", "typeof", "instanceof",
-    "in", "as", "is",
-];
-
-#[rustfmt::skip]
-const DECLARATION_KEYWORDS: &[&str] = &[
-    "const", "let", "var", "val", "type", "fun", "def",
-    "pub", "public", "private", "protected", "internal", "static", "final",
-    "abstract", "async", "export", "import", "mut", "ref", "override",
-    "virtual", "sealed", "lazy", "unsafe", "inline",
-    "suspend", "defer", "go",
-    "declare", "typeset", "local", "read", "readonly", "unset",
-];
-
-/// Pre-scan a token slice to classify each token for spacing decisions.
-///
-/// Skips `$`-prefixed interpolation markers (their contents are Rust
-/// expressions, not target-language tokens).
-fn annotate_tokens(tokens: &[TokenTree], lang: MacroLang) -> Vec<TokenAnnotation> {
-    let mut annotations = vec![TokenAnnotation::Normal; tokens.len()];
-    let mut generic_stack: Vec<usize> = Vec::new();
-    let mut i = 0;
-
-    while i < tokens.len() {
-        let tt = &tokens[i];
-
-        // Skip $ interpolation markers — mirrors the main pass logic.
-        if let TokenTree::Punct(p) = tt
-            && p.as_char() == '$'
-        {
-            i += 1;
-            if i >= tokens.len() {
-                break;
-            }
-            let next = &tokens[i];
-            // $$ or $> or $< or $+ — skip one more token
-            if let TokenTree::Punct(p2) = next
-                && matches!(p2.as_char(), '$' | '>' | '<' | '+')
-            {
-                i += 1;
-                continue;
-            }
-            // $W — skip one ident
-            if is_ident(next, "W") {
-                i += 1;
-                continue;
-            }
-            // $join(expr) or $T(expr) etc — skip ident + group
-            if let TokenTree::Ident(id) = next {
-                let is_type_interp = *id == "T";
-                i += 1;
-                if i < tokens.len() && matches!(&tokens[i], TokenTree::Group(_)) {
-                    i += 1;
-                }
-                // $T(...) always produces a type — mark following `<` as generic
-                // but NOT if it's `<<` (shift operator)
-                if is_type_interp
-                    && i < tokens.len()
-                    && let TokenTree::Punct(p) = &tokens[i]
-                    && p.as_char() == '<'
-                {
-                    // Check if this is `<<` (shift) by looking at spacing
-                    let is_shift = p.spacing() == Spacing::Joint
-                        && i + 1 < tokens.len()
-                        && matches!(&tokens[i + 1], TokenTree::Punct(np) if np.as_char() == '<');
-                    if !is_shift {
-                        annotations[i] = TokenAnnotation::GenericOpen;
-                        generic_stack.push(i);
-                    }
-                }
-                continue;
-            }
-            continue;
-        }
-
-        match tt {
-            TokenTree::Punct(p) => {
-                let ch = p.as_char();
-                match ch {
-                    ':' => {
-                        // PathSepComplete: first `:` is Joint, next is `:`, and
-                        // the `::` is span-adjacent to the preceding token
-                        // (no whitespace before `::` → path separator like `std::fmt`).
-                        // When user writes `fmap :: Type` with space, it's an operator.
-                        if p.spacing() == Spacing::Joint
-                            && i + 1 < tokens.len()
-                            && let TokenTree::Punct(next_p) = &tokens[i + 1]
-                            && next_p.as_char() == ':'
-                        {
-                            let is_path_sep = i > 0 && {
-                                let prev_end = tokens[i - 1].span().end();
-                                let colon_start = p.span().start();
-                                prev_end.line == colon_start.line
-                                    && prev_end.column == colon_start.column
-                            };
-                            if is_path_sep {
-                                annotations[i + 1] = TokenAnnotation::PathSepComplete;
-                            } else {
-                                annotations[i] = TokenAnnotation::DoubleColonOp;
-                            }
-                        }
-                        // MethodCallColon: single `:` span-adjacent to both neighbors
-                        // (Lua `obj:method()`)
-                        else if p.spacing() == Spacing::Alone && i > 0 && i + 1 < tokens.len() {
-                            let prev_end = tokens[i - 1].span().end();
-                            let colon_start = p.span().start();
-                            let colon_end = p.span().end();
-                            let next_start = tokens[i + 1].span().start();
-                            if prev_end.line == colon_start.line
-                                && prev_end.column == colon_start.column
-                                && colon_end.line == next_start.line
-                                && colon_end.column == next_start.column
-                            {
-                                annotations[i] = TokenAnnotation::MethodCallColon;
-                            }
-                        }
-                    }
-                    '!' if p.spacing() == Spacing::Alone
-                        && i > 0
-                        && matches!(&tokens[i - 1], TokenTree::Ident(_)) =>
-                    {
-                        annotations[i] = TokenAnnotation::MacroBang;
-                    }
-                    '+' | '-'
-                        if p.spacing() == Spacing::Joint
-                            && i + 1 < tokens.len()
-                            && matches!(&tokens[i + 1], TokenTree::Punct(np) if np.as_char() == ch) =>
-                    {
-                        // ++ or -- : check if postfix (preceded by ident, literal, or group close)
-                        let is_postfix = if i == 0 {
-                            false
-                        } else {
-                            match &tokens[i - 1] {
-                                TokenTree::Ident(id) => {
-                                    let s = id.to_string();
-                                    !CONTROL_FLOW_KEYWORDS.contains(&s.as_str())
-                                }
-                                TokenTree::Literal(_) => true,
-                                TokenTree::Group(g) => matches!(
-                                    g.delimiter(),
-                                    Delimiter::Parenthesis | Delimiter::Bracket
-                                ),
-                                _ => false,
-                            }
-                        };
-                        if is_postfix {
-                            // Look-ahead: if the token after `++`/`--` is an operand
-                            // (ident, literal, group, or `$` interpolation), this is a
-                            // binary operator (e.g. Haskell `++`), not postfix inc/dec.
-                            let after_second = i + 2;
-                            let followed_by_operand = if after_second < tokens.len() {
-                                match &tokens[after_second] {
-                                    TokenTree::Ident(_)
-                                    | TokenTree::Literal(_)
-                                    | TokenTree::Group(_) => true,
-                                    TokenTree::Punct(p2) => p2.as_char() == '$',
-                                }
-                            } else {
-                                false
-                            };
-                            if !followed_by_operand {
-                                annotations[i] = TokenAnnotation::PostfixIncDec;
-                            }
-                        }
-                    }
-                    '&' | '*' | '-' => {
-                        // ArrowOp: `-` that forms `->` and is span-adjacent to
-                        // preceding token (member access like `cfg->host`).
-                        if ch == '-'
-                            && p.spacing() == Spacing::Joint
-                            && i + 1 < tokens.len()
-                            && matches!(&tokens[i + 1], TokenTree::Punct(np) if np.as_char() == '>')
-                            && i > 0
-                        {
-                            let prev_end = tokens[i - 1].span().end();
-                            let cur_start = p.span().start();
-                            if prev_end.line == cur_start.line
-                                && prev_end.column == cur_start.column
-                            {
-                                annotations[i] = TokenAnnotation::ArrowOp;
-                                annotations[i + 1] = TokenAnnotation::ArrowOp;
-                                i += 2;
-                                continue;
-                            }
-                        }
-
-                        // `<-` compound operator: when `-` follows Joint `<` that
-                        // isn't GenericOpen, the pair forms a binary operator (Haskell
-                        // bind, Go channel). Don't mark as PrefixOp.
-                        if ch == '-'
-                            && i > 0
-                            && let TokenTree::Punct(prev_p) = &tokens[i - 1]
-                            && prev_p.as_char() == '<'
-                            && prev_p.spacing() == Spacing::Joint
-                            && annotations[i - 1] != TokenAnnotation::GenericOpen
-                        {
-                            // Go: `<-ch` is prefix receive — suppress space after
-                            // if span-adjacent to next token. `ch <- 42` (send) has
-                            // whitespace after `-`, so it keeps the space.
-                            if lang == MacroLang::GoLang && i + 1 < tokens.len() {
-                                let dash_end = p.span().end();
-                                let next_start = tokens[i + 1].span().start();
-                                if dash_end.line == next_start.line
-                                    && dash_end.column == next_start.column
-                                {
-                                    annotations[i] = TokenAnnotation::PrefixOp;
-                                }
-                            }
-                            i += 1;
-                            continue;
-                        }
-
-                        // DashSep: `-` between two ident/literals with no whitespace
-                        // on either side (hyphenated word like `from-oci-layout`).
-                        // Requires prev to be Ident/Literal to avoid false positives
-                        // (e.g. `-- file` where prev `-` is Punct).
-                        if ch == '-'
-                            && p.spacing() == Spacing::Alone
-                            && i > 0
-                            && i + 1 < tokens.len()
-                            && matches!(&tokens[i - 1], TokenTree::Ident(_) | TokenTree::Literal(_))
-                            && matches!(&tokens[i + 1], TokenTree::Ident(_) | TokenTree::Literal(_))
-                        {
-                            let prev_end = tokens[i - 1].span().end();
-                            let dash_start = p.span().start();
-                            let dash_end = p.span().end();
-                            let next_start = tokens[i + 1].span().start();
-                            if prev_end.line == dash_start.line
-                                && prev_end.column == dash_start.column
-                                && dash_end.line == next_start.line
-                                && dash_end.column == next_start.column
-                            {
-                                annotations[i] = TokenAnnotation::DashSep;
-                                i += 1;
-                                continue;
-                            }
-                        }
-
-                        // PostfixStar: `*` span-adjacent to preceding ident (pointer type like `Config*`).
-                        if ch == '*' && i > 0 && matches!(&tokens[i - 1], TokenTree::Ident(_)) {
-                            let prev_end = tokens[i - 1].span().end();
-                            let star_start = p.span().start();
-                            if prev_end.line == star_start.line
-                                && prev_end.column == star_start.column
-                            {
-                                annotations[i] = TokenAnnotation::PostfixStar;
-                                i += 1;
-                                continue;
-                            }
-                        }
-
-                        // PostfixAmpersand: `&` span-adjacent to preceding ident/keyword
-                        // (reference type like `auto&`, `int&`).
-                        if ch == '&' && i > 0 && matches!(&tokens[i - 1], TokenTree::Ident(_)) {
-                            let prev_end = tokens[i - 1].span().end();
-                            let amp_start = p.span().start();
-                            if prev_end.line == amp_start.line
-                                && prev_end.column == amp_start.column
-                            {
-                                annotations[i] = TokenAnnotation::PostfixStar;
-                                i += 1;
-                                continue;
-                            }
-                        }
-
-                        // PrefixOp: NOT preceded by non-keyword ident, literal, `)`, or `]`
-                        // After keywords like `return`, `-` is prefix (unary minus)
-                        let is_prefix = if i == 0 {
-                            true
-                        } else {
-                            let prev = &tokens[i - 1];
-                            match prev {
-                                TokenTree::Ident(id) => {
-                                    // After keyword → prefix; after variable → binary
-                                    let s = id.to_string();
-                                    CONTROL_FLOW_KEYWORDS.contains(&s.as_str())
-                                        || DECLARATION_KEYWORDS.contains(&s.as_str())
-                                }
-                                TokenTree::Literal(_) => false,
-                                TokenTree::Group(g) => !matches!(
-                                    g.delimiter(),
-                                    Delimiter::Parenthesis | Delimiter::Bracket
-                                ),
-                                TokenTree::Punct(pp) => !matches!(pp.as_char(), ')' | ']'),
-                            }
-                        };
-                        if is_prefix {
-                            annotations[i] = TokenAnnotation::PrefixOp;
-                        }
-
-                        // Shell: `-- file` separator. The second `-` of `--` gets
-                        // PrefixOp above (prev is Punct '-'). In shell mode, if
-                        // NOT span-adjacent to next ident, downgrade to Normal so
-                        // the space is preserved (separator, not flag prefix).
-                        if lang.is_shell()
-                            && ch == '-'
-                            && annotations[i] == TokenAnnotation::PrefixOp
-                            && i > 0
-                            && matches!(&tokens[i - 1], TokenTree::Punct(pp) if pp.as_char() == '-')
-                            && i + 1 < tokens.len()
-                            && matches!(&tokens[i + 1], TokenTree::Ident(_) | TokenTree::Literal(_))
-                        {
-                            let dash_end = p.span().end();
-                            let next_start = tokens[i + 1].span().start();
-                            if !(dash_end.line == next_start.line
-                                && dash_end.column == next_start.column)
-                            {
-                                annotations[i] = TokenAnnotation::Normal;
-                            }
-                        }
-
-                        // DashFlag: standalone `-` span-adjacent to following ident/literal.
-                        // Overrides PrefixOp or Normal with DashFlag to suppress space after.
-                        // Guards: must not be preceded by another `-` (avoids `-- file`
-                        // false positive where proc_macro2 spans are unreliable).
-                        if ch == '-'
-                            && p.spacing() == Spacing::Alone
-                            && i + 1 < tokens.len()
-                            && matches!(&tokens[i + 1], TokenTree::Ident(_) | TokenTree::Literal(_))
-                            && !(i > 0
-                                && matches!(
-                                    &tokens[i - 1],
-                                    TokenTree::Punct(pp) if pp.as_char() == '-'
-                                ))
-                        {
-                            let dash_end = p.span().end();
-                            let next_start = tokens[i + 1].span().start();
-                            if dash_end.line == next_start.line
-                                && dash_end.column == next_start.column
-                            {
-                                annotations[i] = TokenAnnotation::DashFlag;
-                                i += 1;
-                                continue;
-                            }
-                        }
-                    }
-                    '/' if p.spacing() == Spacing::Alone && i + 1 < tokens.len() => {
-                        let slash_end = p.span().end();
-                        let next_start = tokens[i + 1].span().start();
-                        let next_adj = slash_end.line == next_start.line
-                            && slash_end.column == next_start.column;
-
-                        if i > 0 {
-                            // SlashSep: `/` span-adjacent to both neighbors (path like `linux/amd64`).
-                            let prev_end = tokens[i - 1].span().end();
-                            let slash_start = p.span().start();
-                            if prev_end.line == slash_start.line
-                                && prev_end.column == slash_start.column
-                                && next_adj
-                            {
-                                annotations[i] = TokenAnnotation::SlashSep;
-                            }
-                        }
-                        // Shell: leading `/` or non-adjacent prev — treat as path prefix
-                        if lang.is_shell()
-                            && annotations[i] != TokenAnnotation::SlashSep
-                            && next_adj
-                        {
-                            annotations[i] = TokenAnnotation::SlashSep;
-                        }
-                    }
-                    '.' if lang.is_shell() => {
-                        // Shell: `.` not span-adjacent to prev is a standalone argument
-                        // (e.g. `find .`, `cd ..`), not member access. Handles both
-                        // Alone (single `.`) and Joint (first `.` of `..`).
-                        // Guard: if the dot IS span-adjacent to the next non-dot token,
-                        // it's a dotfile prefix (`.gitignore`) — keep as Normal.
-                        let not_adj_to_prev = if i > 0 {
-                            let prev_end = tokens[i - 1].span().end();
-                            let dot_start = p.span().start();
-                            !(prev_end.line == dot_start.line
-                                && prev_end.column == dot_start.column)
-                        } else {
-                            true
-                        };
-
-                        if not_adj_to_prev {
-                            // Check if this dot (or `..` sequence) is adjacent to
-                            // the following non-dot token — if so, it's a dotfile.
-                            let seq_end = if p.spacing() == Spacing::Joint
-                                && i + 1 < tokens.len()
-                                && matches!(&tokens[i + 1], TokenTree::Punct(p2) if p2.as_char() == '.')
-                            {
-                                i + 2 // `..` — check token after second dot
-                            } else {
-                                i + 1 // single `.` — check next token
-                            };
-
-                            let adj_to_next = if seq_end < tokens.len() {
-                                let dot_seq_end = tokens[seq_end - 1].span().end();
-                                let next_start = tokens[seq_end].span().start();
-                                dot_seq_end.line == next_start.line
-                                    && dot_seq_end.column == next_start.column
-                            } else {
-                                false
-                            };
-
-                            if !adj_to_next {
-                                annotations[i] = TokenAnnotation::DotArg;
-                                // If Joint (first of `..`), also mark the second dot
-                                if seq_end == i + 2 {
-                                    annotations[i + 1] = TokenAnnotation::DotArg;
-                                }
-                            }
-                        }
-                    }
-                    '?' => {
-                        // SafeCallQ: Joint `?` followed by `.`
-                        if p.spacing() == Spacing::Joint
-                            && i + 1 < tokens.len()
-                            && let TokenTree::Punct(next_p) = &tokens[i + 1]
-                            && next_p.as_char() == '.'
-                        {
-                            annotations[i] = TokenAnnotation::SafeCallQ;
-                        }
-                        // PostfixQuestion: `?` span-adjacent to preceding ident or group-close
-                        // (e.g. `Int?`, `String?`, `(Int)?`)
-                        else if i > 0 {
-                            let prev_end = tokens[i - 1].span().end();
-                            let q_start = p.span().start();
-                            let is_adjacent =
-                                prev_end.line == q_start.line && prev_end.column == q_start.column;
-                            if is_adjacent {
-                                let is_type_context = match &tokens[i - 1] {
-                                    TokenTree::Ident(_) => true,
-                                    TokenTree::Group(g) => matches!(
-                                        g.delimiter(),
-                                        Delimiter::Parenthesis | Delimiter::Bracket
-                                    ),
-                                    _ => false,
-                                };
-                                if is_type_context {
-                                    annotations[i] = TokenAnnotation::PostfixQuestion;
-                                }
-                            }
-                        }
-                    }
-                    '<' => {
-                        // GenericOpen: preceded by uppercase Ident, PathSepComplete,
-                        // or Joint `:` (turbofish).
-                        // Joint `<` followed by another `<` is `<<` (shift), not generic.
-                        let is_shift = p.spacing() == Spacing::Joint
-                            && i + 1 < tokens.len()
-                            && matches!(&tokens[i + 1], TokenTree::Punct(np) if np.as_char() == '<' || np.as_char() == '=');
-                        let is_generic = if is_shift || i == 0 {
-                            false
-                        } else {
-                            let prev = &tokens[i - 1];
-                            match prev {
-                                TokenTree::Ident(id) => {
-                                    let s = id.to_string();
-                                    // Uppercase ident (type name heuristic), OR
-                                    // declaration keyword (e.g. `public <T>`), OR
-                                    // any ident preceded by PathSepComplete
-                                    // (e.g., `std::map<` — lowercase but qualified), OR
-                                    // span-adjacent lowercase ident (`identity<T>`)
-                                    s.starts_with(|c: char| c.is_uppercase())
-                                        || DECLARATION_KEYWORDS.contains(&s.as_str())
-                                        || (i >= 2
-                                            && annotations[i - 2]
-                                                == TokenAnnotation::PathSepComplete)
-                                        || {
-                                            let prev_end = id.span().end();
-                                            let lt_start = p.span().start();
-                                            prev_end.line == lt_start.line
-                                                && prev_end.column == lt_start.column
-                                        }
-                                }
-                                TokenTree::Punct(pp) => {
-                                    // After PathSepComplete or Joint `:` (turbofish)
-                                    (pp.as_char() == ':'
-                                        && annotations[i - 1] == TokenAnnotation::PathSepComplete)
-                                        || (pp.as_char() == ':' && pp.spacing() == Spacing::Joint)
-                                }
-                                _ => false,
-                            }
-                        };
-                        if is_generic {
-                            annotations[i] = TokenAnnotation::GenericOpen;
-                            generic_stack.push(i);
-                        }
-                    }
-                    '>' if !generic_stack.is_empty() => {
-                        generic_stack.pop();
-                        annotations[i] = TokenAnnotation::GenericClose;
-                    }
-                    ';' => {
-                        // Reset generic stack at statement boundaries
-                        generic_stack.clear();
-                    }
-                    '=' if i > 0 => {
-                        let prev_is_assignable = matches!(
-                            &tokens[i - 1],
-                            TokenTree::Ident(_) | TokenTree::Literal(_) | TokenTree::Group(_)
-                        );
-                        if prev_is_assignable {
-                            let prev_end = tokens[i - 1].span().end();
-                            let eq_start = p.span().start();
-                            if prev_end.line == eq_start.line && prev_end.column == eq_start.column
-                            {
-                                // Don't mark as AssignAdjacent if this is part of
-                                // ==, =>, <=, >= (Joint followed by = or >)
-                                let is_compound = p.spacing() == Spacing::Joint
-                                    && i + 1 < tokens.len()
-                                    && matches!(&tokens[i + 1], TokenTree::Punct(np) if np.as_char() == '=' || np.as_char() == '>');
-                                if !is_compound {
-                                    annotations[i] = TokenAnnotation::AssignAdjacent;
-                                }
-                            }
-                        }
-                    }
-                    _ => {}
-                }
-            }
-            TokenTree::Group(g)
-                if i > 0
-                    && matches!(g.delimiter(), Delimiter::Parenthesis | Delimiter::Bracket) =>
-            {
-                // Mark groups that are span-adjacent to the preceding ident/literal/>
-                // as CallOpen (function call / array index). Non-adjacent groups
-                // or groups after operators/keywords get default spacing.
-                let prev_is_callable = match &tokens[i - 1] {
-                    TokenTree::Ident(id) => {
-                        let s = id.to_string();
-                        let is_keyword = CONTROL_FLOW_KEYWORDS.contains(&s.as_str())
-                            || DECLARATION_KEYWORDS.contains(&s.as_str());
-                        if !is_keyword {
-                            true
-                        } else if i >= 2 {
-                            matches!(&tokens[i - 2], TokenTree::Punct(p) if p.as_char() == '.')
-                        } else {
-                            false
-                        }
-                    }
-                    TokenTree::Literal(_) | TokenTree::Group(_) => true,
-                    TokenTree::Punct(p) => p.as_char() == '>',
-                };
-                if prev_is_callable {
-                    let prev_end = tokens[i - 1].span().end();
-                    let group_start = g.span().start();
-                    if prev_end.line == group_start.line && prev_end.column == group_start.column {
-                        annotations[i] = TokenAnnotation::CallOpen;
-                    }
-                }
-            }
-            _ => {}
-        }
-        i += 1;
-    }
-
-    annotations
-}
 
 /// Convert a sequence of tokens into a format string and typed argument list.
 ///
@@ -1127,273 +469,4 @@ pub(super) fn split_join_args(
     }
 
     Ok((sep_tokens, iter_tokens))
-}
-
-/// Insert a space between the previous and current tokens if needed.
-pub(super) fn maybe_space(
-    format: &mut String,
-    state: &SpacingState,
-    current: PrevTokenKind,
-    annotation: TokenAnnotation,
-) {
-    let prev = state.prev;
-
-    if prev == PrevTokenKind::None || prev == PrevTokenKind::GroupOpen {
-        return;
-    }
-
-    // %W already provides a space (or newline), so don't add another.
-    if prev == PrevTokenKind::SoftBreak {
-        return;
-    }
-
-    // Annotation-based suppression (replaces old suppress_space flag).
-    match annotation {
-        TokenAnnotation::MacroBang
-        | TokenAnnotation::GenericClose
-        | TokenAnnotation::SafeCallQ
-        | TokenAnnotation::PostfixIncDec
-        | TokenAnnotation::PostfixStar
-        | TokenAnnotation::PostfixQuestion
-        | TokenAnnotation::ArrowOp
-        | TokenAnnotation::AssignAdjacent
-        | TokenAnnotation::MethodCallColon
-        | TokenAnnotation::DashSep
-        | TokenAnnotation::SlashSep => return,
-        TokenAnnotation::GenericOpen if prev != PrevTokenKind::Keyword => return,
-        _ => {}
-    }
-
-    // No space after prefix operators, path separators, or generic openers.
-    if matches!(
-        prev,
-        PrevTokenKind::PrefixOp(_)
-            | PrevTokenKind::PathSep
-            | PrevTokenKind::GenericOpen
-            | PrevTokenKind::DollarLiteral
-    ) {
-        return;
-    }
-
-    // No space before certain punctuation.
-    if let PrevTokenKind::Punct(ch, _) = current {
-        match ch {
-            ',' | ';' | ')' | ']' => return,
-            '.' if annotation != TokenAnnotation::DotArg => return,
-            ':' if annotation != TokenAnnotation::DoubleColonOp => match state.colon_ctx {
-                ColonContext::Ternary | ColonContext::WalrusAssign | ColonContext::ForRange => {}
-                ColonContext::TypeAnnotation
-                | ColonContext::MapEntry
-                | ColonContext::PathSeparator => return,
-            },
-            _ => {}
-        }
-    }
-
-    // No space between joint punctuation (===, !==, ->, ::, etc.).
-    if let PrevTokenKind::Punct(_, Spacing::Joint) = prev {
-        return;
-    }
-
-    // No space after opening punctuation.
-    if let PrevTokenKind::Punct('(' | '[' | '.' | '!' | '~' | '@' | '#', _) = prev {
-        return;
-    }
-
-    // No space before `(` or `[` when span-adjacent to preceding token
-    // (function call, array index). Non-adjacent groups get default spacing.
-    if annotation == TokenAnnotation::CallOpen {
-        return;
-    }
-
-    // No space before a group when preceded by a specifier ($T, $N, etc.)
-    // — handles `$T(x){}` struct literals and `$T(x)()` calls.
-    if let PrevTokenKind::GroupOpen = current
-        && prev == PrevTokenKind::Specifier
-    {
-        return;
-    }
-
-    // Default: add a space between tokens.
-    format.push(' ');
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use proc_macro2::{TokenStream, TokenTree};
-
-    fn annotate_src(src: &str) -> Vec<(String, TokenAnnotation)> {
-        let ts: TokenStream = src.parse().unwrap();
-        let tokens: Vec<TokenTree> = ts.into_iter().collect();
-        let annotations = annotate_tokens(&tokens, MacroLang::Unaware);
-        tokens
-            .iter()
-            .zip(annotations.iter())
-            .map(|(tt, ann)| (tt.to_string(), *ann))
-            .collect()
-    }
-
-    fn annotate_lang(src: &str, lang: MacroLang) -> Vec<(String, TokenAnnotation)> {
-        let ts: TokenStream = src.parse().unwrap();
-        let tokens: Vec<TokenTree> = ts.into_iter().collect();
-        let annotations = annotate_tokens(&tokens, lang);
-        tokens
-            .iter()
-            .zip(annotations.iter())
-            .map(|(tt, ann)| (tt.to_string(), *ann))
-            .collect()
-    }
-
-    fn ann_at(anns: &[(String, TokenAnnotation)], idx: usize) -> TokenAnnotation {
-        anns[idx].1
-    }
-
-    #[test]
-    fn generic_open_close() {
-        let a = annotate_src("Vec < T >");
-        assert_eq!(ann_at(&a, 1), TokenAnnotation::GenericOpen);
-        assert_eq!(ann_at(&a, 3), TokenAnnotation::GenericClose);
-    }
-
-    #[test]
-    fn generic_not_triggered_for_lowercase() {
-        let a = annotate_src("x < y >");
-        assert_eq!(ann_at(&a, 1), TokenAnnotation::Normal);
-    }
-
-    #[test]
-    fn macro_bang() {
-        let a = annotate_src("println !");
-        assert_eq!(ann_at(&a, 1), TokenAnnotation::MacroBang);
-    }
-
-    #[test]
-    fn path_sep_adjacent() {
-        let a = annotate_src("std::fmt");
-        // tokens: "std" ":" ":" "fmt"
-        // Second ":" gets MethodCallColon (adjacent both sides); both
-        // MethodCallColon and PathSepComplete suppress space — same effect.
-        assert_eq!(ann_at(&a, 2), TokenAnnotation::MethodCallColon);
-    }
-
-    #[test]
-    fn double_colon_op_spaced() {
-        // fmap :: Type — space before `::` makes it an operator
-        let a = annotate_src("fmap :: Type");
-        assert_eq!(ann_at(&a, 1), TokenAnnotation::DoubleColonOp);
-    }
-
-    #[test]
-    fn call_open_adjacent() {
-        // f(x) — group adjacent to ident → CallOpen
-        let a = annotate_src("f(x)");
-        assert_eq!(ann_at(&a, 1), TokenAnnotation::CallOpen);
-    }
-
-    #[test]
-    fn call_open_not_spaced() {
-        // f (x) — space before paren → Normal
-        let a = annotate_src("f (x)");
-        assert_eq!(ann_at(&a, 1), TokenAnnotation::Normal);
-    }
-
-    #[test]
-    fn dash_sep_hyphenated() {
-        let a = annotate_src("from-oci-layout");
-        // "from" "-" "oci" "-" "layout"
-        assert_eq!(ann_at(&a, 1), TokenAnnotation::DashSep);
-        assert_eq!(ann_at(&a, 3), TokenAnnotation::DashSep);
-    }
-
-    #[test]
-    fn slash_sep_path() {
-        let a = annotate_src("linux/amd64");
-        // "linux" "/" "amd64"
-        assert_eq!(ann_at(&a, 1), TokenAnnotation::SlashSep);
-    }
-
-    #[test]
-    fn arrow_op_adjacent() {
-        let a = annotate_src("ptr->field");
-        // "ptr" "-" ">" "field"
-        assert_eq!(ann_at(&a, 1), TokenAnnotation::ArrowOp);
-    }
-
-    #[test]
-    fn prefix_op_star() {
-        let a = annotate_src("= *ptr");
-        // "=" "*" "ptr" — * after = is prefix
-        assert_eq!(ann_at(&a, 1), TokenAnnotation::PrefixOp);
-    }
-
-    #[test]
-    fn prefix_op_ampersand() {
-        let a = annotate_src("= &x");
-        assert_eq!(ann_at(&a, 1), TokenAnnotation::PrefixOp);
-    }
-
-    #[test]
-    fn postfix_question_adjacent() {
-        let a = annotate_src("Int?");
-        // "Int" "?" — adjacent → PostfixQuestion
-        assert_eq!(ann_at(&a, 1), TokenAnnotation::PostfixQuestion);
-    }
-
-    #[test]
-    fn postfix_star_adjacent() {
-        let a = annotate_src("Config*");
-        // "Config" "*" — adjacent → PostfixStar
-        assert_eq!(ann_at(&a, 1), TokenAnnotation::PostfixStar);
-    }
-
-    #[test]
-    fn assign_adjacent() {
-        let a = annotate_src("NAME=val");
-        // "NAME" "=" "val"
-        assert_eq!(ann_at(&a, 1), TokenAnnotation::AssignAdjacent);
-    }
-
-    // --- Language-specific ---
-
-    #[test]
-    fn bash_dash_flag() {
-        let a = annotate_lang("-avz", MacroLang::Bash);
-        assert_eq!(ann_at(&a, 0), TokenAnnotation::DashFlag);
-    }
-
-    #[test]
-    fn bash_dot_arg() {
-        let a = annotate_lang("find .", MacroLang::Bash);
-        assert_eq!(ann_at(&a, 1), TokenAnnotation::DotArg);
-    }
-
-    #[test]
-    fn bash_slash_leading_path() {
-        let a = annotate_lang("/usr/local", MacroLang::Bash);
-        // "/" "usr" "/" "local"
-        assert_eq!(ann_at(&a, 0), TokenAnnotation::SlashSep);
-    }
-
-    #[test]
-    fn go_receive_adjacent() {
-        // <-ch → < is Joint, - adjacent to ch → PrefixOp on -
-        let a = annotate_lang("<-ch", MacroLang::GoLang);
-        assert_eq!(ann_at(&a, 1), TokenAnnotation::PrefixOp);
-    }
-
-    #[test]
-    fn go_send_not_prefix() {
-        // ch <- 42 → - is NOT adjacent to 42 → stays Normal
-        let a = annotate_lang("ch <- 42", MacroLang::GoLang);
-        assert_eq!(ann_at(&a, 2), TokenAnnotation::Normal);
-    }
-
-    #[test]
-    fn normal_tokens_stay_normal() {
-        let a = annotate_src("let x = 42");
-        for (_, ann) in &a {
-            assert_eq!(*ann, TokenAnnotation::Normal);
-        }
-    }
 }
